refactor structure of flask server
