+++ app/api/__init__.py
@@ -1,0 +1,1 @@
+# Package initialization for api module
